+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v1.1.0
 
 ### Added
-- Added `sectionCustomizationBySection` schema and resolver
+- Added `sectionCustomizationByVersionedSection` and `sectionCustomizationByVersionedQuestion` schema and resolver, and added `findByCustomizationAndVersionedSection` and `findByCustomizationAndVersionedQuestion` methods to the `SectionCustomization` model.
 - Added `SectionCustomization` and `QuestionCustomization` schemas and resolvers
 - Added a new `authenticatedResolver` wrapper function to help handle common authorization checks on a resolver
 - Added `TemplateCustomizationOverview` to the `TemplateCustomization` model which returns high level overview info about the base funder template and the customizations